requirements.txt
Django==2.1.4
django-crispy-forms==1.7.2
Pillow==5.3.0
pkg-resources==0.0.0
pycrypto==2.6.1
pytz==2018.7
requests==2.21.0
